sci-libs/trilinos: Fix install
Closes: https://bugs.gentoo.org/691888
Package-Manager: Portage-2.3.71, Repoman-2.3.17

+++ sci-libs/trilinos/trilinos-12.14.1.ebuild
@@ -195,7 +195,9 @@ src_install() {
 
 	# Clean up the mess:
 	mv "${ED}"/bin "${ED}/usr/$(get_libdir)"/trilinos || die "mv failed"
-	mv "${ED}"/lib/exodus.py "${ED}/usr/$(get_libdir)"/trilinos || die "mv failed"
+	if [ -f "${ED}"/lib/exodus.py ]; then
+		mv "${ED}"/lib/exodus.py "${ED}/usr/$(get_libdir)"/trilinos || die "mv failed"
+	fi
 
 	#
 	# register $(get_libdir)/trilinos in LDPATH so that the dynamic linker
